requirements.txt
streamlit
google-generativeai
python-dotenv
pandas
Pillow
PyPDF2
